Refactoring LinearTrainer to share code.
